+++ CHANGELOG.md
@@ -31,6 +31,7 @@ Version counting is based on semantic versioning (Major.Feature.Patch)
 
 ### WebUI
 * Add a very basic web reader.
+* Add rescan xml functionality to the webui.
 
 ### All apps
 * Unify comics sorting to use the string based universal number everywhere.